+++ tvshows/src/components/features/review/ShowReviewSection/ShowReviewSection.tsx
@@ -1,22 +1,34 @@
 "use client";
-import { IReviewList } from "@/typings/review.types";
+import { fetcher } from "@/fetchers/fetcher";
+import { swrKeys } from "@/fetchers/swrKeys";
+import { useParams } from "next/navigation";
+import useSWR from "swr";
 import { ReviewForm } from "../ReviewForm/ReviewForm";
 import { ReviewList } from "../ReviewList/ReviewList";
 
-interface IShowReviewSectionProps {
-  showId: number;
-  reviewList: IReviewList;
-}
+export const ShowReviewSection = () => {
+  const params = useParams();
+  const showId = params?.id;
 
-export const ShowReviewSection = ({
-  showId,
-  reviewList,
-}: IShowReviewSectionProps) => {
+  const {
+    data: reviewListResponse,
+    error: reviewError,
+    isLoading: reviewIsLoading,
+  } = useSWR(swrKeys.reviews(Number(showId)), { fetcher });
+
+  if (reviewIsLoading) {
+    return <div>Loading...</div>;
+  }
+
+  if (reviewError) {
+    return <div>Oops, something went wrong...</div>;
+  }
+
   return (
     <>
-      <ReviewForm showId = {showId} />
-      {reviewList && reviewList.reviews.length > 0 && (
-        <ReviewList reviewList={reviewList} />
+      <ReviewForm showId={Number(showId)} />
+      {reviewListResponse && reviewListResponse.reviews.length > 0 && (
+        <ReviewList reviewList={reviewListResponse} />
       )}
     </>
   );
