+++ tvshows/src/components/features/shows/ShowContainer/ShowContainer.tsx
@@ -1,36 +1,8 @@
 import { ShowReviewSection } from "@/components/features/review/ShowReviewSection/ShowReviewSection";
-import { fetcher } from "@/fetchers/fetcher";
-import { swrKeys } from "@/fetchers/swrKeys";
 import { Container } from "@chakra-ui/react";
-import { useParams } from "next/navigation";
-import useSWR from "swr";
 import { ShowDetailsCard } from "../ShowDetailsCard/ShowDetailsCard";
 
 export const ShowContainer = () => {
-  const params = useParams();
-  const showId = params?.id;
-
-  const {
-    data: showListResponse,
-    error: showError,
-    isLoading: showIsLoading,
-  } = useSWR(swrKeys.show(Number(showId)), { fetcher });
-
-  const {
-    data: reviewListResponse,
-    error: reviewError,
-    isLoading: reviewIsLoading,
-  } = useSWR(swrKeys.reviews(Number(showId)), { fetcher });
-
-
-  if (showIsLoading || reviewIsLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (showError || reviewError) {
-    return <div>Oops, something went wrong...</div>;
-  }
-
   return (
     <Container
       maxW="container.lg"
@@ -41,13 +13,8 @@ export const ShowContainer = () => {
       justifyContent="center"
       textAlign="center"
     >
-      {showListResponse && (
-        <ShowDetailsCard show={showListResponse}/>
-      )}
-      <ShowReviewSection
-        reviewList={reviewListResponse}
-        showId={Number(showId)}
-      />
+      <ShowDetailsCard />
+      <ShowReviewSection />
     </Container>
   );
 };
